refactor(core): `createFhircastEventContext` -> `test-utils.ts`
